fix: ``size_t`` does not hold file sizes.

+++ include/libmcx/mcx.h
@@ -7,6 +7,7 @@
 
 #include <libmcx/atrb.h>
 #include <libmcx/types.h>
+#include <stdio.h>
 
 #define MCX_SECTOR     0x1000 /* Size in bytes for a sector. */
 #define MCX_TABLE      0x1000 /* Size in bytes for one table. */
@@ -14,19 +15,19 @@
 #define MCX_TABLE_LEN  0x400  /* Amount of elements within one table. */
 #define MCX_TABLES_LEN 0x800  /* Amount of elements within the tables. */
 
-void mcx_check(const void *mcx, size_t size, const char *pat);
+void mcx_check(const void *mcx, off_t size, const char *pat);
 
 /* Repairs faults in the file format,
  * outputting warnings describing the repaired fault & affected chunk.
  * Returns the size of the file, which may be larger than,
  * less than, or equal to the original size. */
-size_t mcx_repair(void *mcx, size_t size);
+off_t mcx_repair(void *mcx, off_t size);
 
 /* Prunes the unused sections in a .mcX file.
  * It is assumed that the table is formatted correctly and
  * that all data is accessible.
  * Returns the new file size.*/
-size_t mcx_defrag(void *mcx);
+off_t mcx_defrag(void *mcx);
 
 /* Computes the size in bytes of the entire .mcX file,
  * according to the table.
@@ -37,7 +38,7 @@ size_t mcx_defrag(void *mcx);
  * sectors at the end, or that overlapping values are present.
  * If this value is higher than the actual file size,
  * then corruption has taken place. */
-size_t mcx_calcsize(const void *mcx) PURE;
+off_t mcx_calcsize(const void *mcx) PURE;
 
 /* Computes the minimum size in bytes required for the entire .mcX file,
  * according to the table.
@@ -48,6 +49,6 @@ size_t mcx_calcsize(const void *mcx) PURE;
  * of the actual size due to the presence of unused sectors.
  * In the edge case of overlapping sectors, this value may be larger
  * than the actual file size. */
-size_t mcx_sumsize(const void *mcx) PURE;
+off_t mcx_sumsize(const void *mcx) PURE;
 
 #endif /* MCXEDIT_MCX_H */
